+++ src/main/java/io/github/astrapi69/junit/jupiter/converter/StringToByteArrayConverter.java
@@ -1,0 +1,53 @@
+/**
+ * The MIT License
+ *
+ * Copyright (C) 2021 Asterios Raptis
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining
+ * a copy of this software and associated documentation files (the
+ * "Software"), to deal in the Software without restriction, including
+ * without limitation the rights to use, copy, modify, merge, publish,
+ * distribute, sublicense, and/or sell copies of the Software, and to
+ * permit persons to whom the Software is furnished to do so, subject to
+ * the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be
+ * included in all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+ * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+ * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+ * LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+ * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+ * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+package io.github.astrapi69.junit.jupiter.converter;
+
+import org.junit.jupiter.params.converter.ArgumentConversionException;
+import org.junit.jupiter.params.converter.SimpleArgumentConverter;
+
+/**
+ * Custom argument converter to convert a string to a byte array for parameterized tests.
+ */
+public class StringToByteArrayConverter extends SimpleArgumentConverter
+{
+	/** {@inheritDoc} */
+	@Override
+	protected Object convert(Object source, Class<?> targetType) throws ArgumentConversionException
+	{
+		if (source == null)
+		{
+			throw new ArgumentConversionException("source is null");
+		}
+		if (!(source instanceof String))
+		{
+			throw new ArgumentConversionException("source is not a string");
+		}
+		if (!targetType.equals(byte[].class))
+		{
+			throw new ArgumentConversionException("target type is not byte[]");
+		}
+		return ((String)source).getBytes();
+	}
+}